+++ examples/screenshot-asynchronous.rs
@@ -2,7 +2,6 @@
 extern crate glium;
 extern crate image;
 
-use std::path::Path;
 use std::thread;
 
 use glium::{glutin, Surface};
@@ -249,7 +248,7 @@ fn main() {
 
             match event {
                 Event::WindowEvent { event, .. } => match event {
-                    WindowEvent::Closed => closed = true,
+                    WindowEvent::CloseRequested => closed = true,
                     WindowEvent::KeyboardInput { input, .. } => {
                         if let ElementState::Pressed = input.state {
                             if let Some(VirtualKeyCode::S) = input.virtual_keycode {
@@ -310,9 +309,7 @@ fn main() {
 
                 // Save the screenshot to file
                 let image = image::DynamicImage::ImageRgba8(image_buffer).flipv();
-                let mut output = std::fs::File::create(&Path::new("glium-example-screenshot.png"))
-                    .unwrap();
-                image.save(&mut output, image::ImageFormat::PNG).unwrap();
+                image.save("glium-example-screenshot.png").unwrap();
             });
         }
     }