+++ src/context/mod.rs
@@ -524,6 +524,8 @@ impl Context {
         let dimensions = self.get_framebuffer_dimensions();
         let rect = ::Rect { left: 0, bottom: 0, width: dimensions.0, height: dimensions.1 };
 
+        // FIXME: See https://github.com/glium/glium/pull/1682#issuecomment-375596152
+        // Leave it unchecked for now, it will generate warning as a reminder it needs fixing.
         let mut data = Vec::with_capacity(0);
         ops::read(&mut ctxt, ops::Source::DefaultFramebuffer(gl::FRONT_LEFT), &rect,
                           &mut data, false);